Statisztika osztály kész

+++ Motor.cs
@@ -11,10 +11,10 @@ namespace Motors
         private string brand;
         private string name;
         private int releaseYear;
-        private int performance;
-        private int priceInEur;
+        private double performance;
+        private double priceInEur;
 
-        public Motor(string brand, string name, int releaseYear, int performance, int priceInEur)
+        public Motor(string brand, string name, int releaseYear, double performance, double priceInEur)
         {
             this.brand = brand;
             this.name = name;
@@ -26,8 +26,8 @@ namespace Motors
         public string Brand { get => brand;}
         public string Name { get => name;}
         public int ReleaseYear { get => releaseYear; }
-        public int Performance { get => performance; }
-        public int PriceInEur { get => priceInEur; }
+        public double Performance { get => performance; }
+        public double PriceInEur { get => priceInEur; }
